requirements.txt
psycopg2-binary==2.9.11
fastapi==0.115.11
uvicorn==0.34.0
sqlalchemy==2.0.40
pydantic==2.11.1
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
python-dotenv==1.1.0
authlib==1.5.1
httpx==0.28.1
itsdangerous==2.2.0
pymysql==1.1.1
aiofiles==24.1.0
supabase==2.15.1
